Phase 1 - JUnit Tests For Merge Sort
Copied the JUnit tests for the bubble sort method and refactored them for the merge sort method as I expect both of the sort algorithms to pass the same tests. Made some commenting and syntax changes to the bubble sort method for clarity when reading the code and running tests.