Phase 2 - Implemented MVC & User Input
Implemented the MVC design pattern into the program. Refactored all the code in the SortManagerMain class so that the sort algorithms are within their respective classes. Also any methods todo with the controller or view were moved accordingly. Refactored the JUnit tests so they can access the bubble and merge algorithms they are testing in their new classes. Implemented user input methods into the view and other methods that the main method will call to take user input or display output to the user.