Phase 3 - Implemented Log4j
Added the Log4j dependency into the pom.xml file and implemented a Log4j properties file to specify that we would like the logger to output to a file following termination of the program. Also initialized the logger within SortManager and tested that it works by calling it within the initiateSort method.