Phase 5 - Implemented Timings & Program Loop
Implemented execution timing being output to the console after the execution of a sort algorithm to allow for comparison between them. Implemented a program loop into the controller that can be called in the main and the ability for the user to exit the program or loop it after the use of a sort algorithm. Updated README to reflect phase 5 changes.